Updated files
Added more board management ideas and game logic and ship functions

Some of the challenges will be how to interact with the gui board and
placing pieces but the logic and all that should be set. Been going
through it in my head. Works similar to the way I did tic tac toe any
size board, n x n